Move json parser to misc.
